// src/interfaces/streams.ts
export interface ReadReq {
  options: ReadReq_Options | undefined;
}

export interface ReadReq_Options {
  stream?: ReadReq_Options_StreamOptions | undefined;
  all?: ReadReq_Options_AllOptions | undefined;
  readDirection: ReadReq_Options_ReadDirection;
  resolveLinks: boolean;
  count?: number | undefined;
  subscription?: ReadReq_Options_SubscriptionOptions | undefined;
  filter?: ReadReq_Options_FilterOptions | undefined;
  noFilter?: Empty | undefined;
  uuidOption:
    | ReadReq_Options_UUIDOption
    | undefined;
  /** @deprecated */
  controlOption: ReadReq_Options_ControlOption | undefined;
}

export enum ReadReq_Options_ReadDirection {
  Forwards = 0,
  Backwards = 1,
  UNRECOGNIZED = -1,
}

export interface ReadReq_Options_StreamOptions {
  streamIdentifier: StreamIdentifier | undefined;
  revision?: number | undefined;
  start?: Empty | undefined;
  end?: Empty | undefined;
}

export interface ReadReq_Options_AllOptions {
  position?: ReadReq_Options_Position | undefined;
  start?: Empty | undefined;
  end?: Empty | undefined;
}

export interface ReadReq_Options_SubscriptionOptions {
}

export interface ReadReq_Options_Position {
  commitPosition: number;
  preparePosition: number;
}

export interface ReadReq_Options_FilterOptions {
  streamIdentifier?: ReadReq_Options_FilterOptions_Expression | undefined;
  eventType?: ReadReq_Options_FilterOptions_Expression | undefined;
  max?: number | undefined;
  count?: Empty | undefined;
  checkpointIntervalMultiplier: number;
}

export interface ReadReq_Options_FilterOptions_Expression {
  regex: string;
  prefix: string[];
}

export interface ReadReq_Options_UUIDOption {
  structured?: Empty | undefined;
  string?: Empty | undefined;
}

export interface ReadReq_Options_ControlOption {
  compatibility: number;
}

export interface ReadResp {
  event?: ReadResp_ReadEvent | undefined;
  confirmation?: ReadResp_SubscriptionConfirmation | undefined;
  checkpoint?: ReadResp_Checkpoint | undefined;
  streamNotFound?:
    | ReadResp_StreamNotFound
    | undefined;
  /** @deprecated */
  firstStreamPosition?:
    | number
    | undefined;
  /** @deprecated */
  lastStreamPosition?:
    | number
    | undefined;
  /** @deprecated */
  lastAllStreamPosition?: AllStreamPosition | undefined;
  caughtUp?: ReadResp_CaughtUp | undefined;
  fellBehind?: ReadResp_FellBehind | undefined;
}

/** The $all or stream subscription has caught up and become live. */
export interface ReadResp_CaughtUp {
  /** Current time in the server when the subscription caught up */
  timestamp:
    | Timestamp
    | undefined;
  /**
   * Checkpoint for resuming a stream subscription.
   * For stream subscriptions it is populated unless the stream is empty.
   * For $all subscriptions it is not populated.
   */
  streamRevision?:
    | number
    | undefined;
  /**
   * Checkpoint for resuming a $all subscription.
   * For stream subscriptions it is not populated.
   * For $all subscriptions it is populated unless the database is empty.
   */
  position?: ReadResp_Position | undefined;
}

/** The $all or stream subscription has fallen back into catchup mode and is no longer live. */
export interface ReadResp_FellBehind {
  /** Current time in the server when the subscription fell behind */
  timestamp:
    | Timestamp
    | undefined;
  /**
   * Checkpoint for resuming a stream subscription.
   * For stream subscriptions it is populated unless the stream is empty.
   * For $all subscriptions it is not populated.
   */
  streamRevision?:
    | number
    | undefined;
  /**
   * Checkpoint for resuming a $all subscription.
   * For stream subscriptions it is not populated.
   * For $all subscriptions it is populated unless the database is empty.
   */
  position?: ReadResp_Position | undefined;
}

export interface ReadResp_ReadEvent {
  event: ReadResp_ReadEvent_RecordedEvent | undefined;
  link: ReadResp_ReadEvent_RecordedEvent | undefined;
  commitPosition?: number | undefined;
  noPosition?: Empty | undefined;
}

export interface ReadResp_ReadEvent_RecordedEvent {
  id: UUID | undefined;
  streamIdentifier: StreamIdentifier | undefined;
  streamRevision: number;
  preparePosition: number;
  commitPosition: number;
  metadata: { [key: string]: string };
  customMetadata: Uint8Array;
  data: Uint8Array;
}

export interface ReadResp_ReadEvent_RecordedEvent_MetadataEntry {
  key: string;
  value: string;
}

export interface ReadResp_SubscriptionConfirmation {
  subscriptionId: string;
}

export interface ReadResp_Checkpoint {
  commitPosition: number;
  preparePosition: number;
  /** Current time in the server when the checkpoint was reached */
  timestamp: Timestamp | undefined;
}

export interface ReadResp_Position {
  commitPosition: number;
  preparePosition: number;
}

export interface ReadResp_StreamNotFound {
  streamIdentifier: StreamIdentifier | undefined;
}

export interface AppendReq {
  options?: AppendReq_Options | undefined;
  proposedMessage?: AppendReq_ProposedMessage | undefined;
}

export interface AppendReq_Options {
  streamIdentifier: StreamIdentifier | undefined;
  revision?: number | undefined;
  noStream?: Empty | undefined;
  any?: Empty | undefined;
  streamExists?: Empty | undefined;
}

export interface AppendReq_ProposedMessage {
  id: UUID | undefined;
  metadata: { [key: string]: string };
  customMetadata: Uint8Array;
  data: Uint8Array;
}

export interface AppendReq_ProposedMessage_MetadataEntry {
  key: string;
  value: string;
}

export interface AppendResp {
  success?: AppendResp_Success | undefined;
  wrongExpectedVersion?: AppendResp_WrongExpectedVersion | undefined;
}

export interface AppendResp_Position {
  commitPosition: number;
  preparePosition: number;
}

export interface AppendResp_Success {
  currentRevision?: number | undefined;
  noStream?: Empty | undefined;
  position?: AppendResp_Position | undefined;
  noPosition?: Empty | undefined;
}

export interface AppendResp_WrongExpectedVersion {
  /** @deprecated */
  currentRevision2060?:
    | number
    | undefined;
  /** @deprecated */
  noStream2060?:
    | Empty
    | undefined;
  /** @deprecated */
  expectedRevision2060?:
    | number
    | undefined;
  /** @deprecated */
  any2060?:
    | Empty
    | undefined;
  /** @deprecated */
  streamExists2060?: Empty | undefined;
  currentRevision?: number | undefined;
  currentNoStream?: Empty | undefined;
  expectedRevision?: number | undefined;
  expectedAny?: Empty | undefined;
  expectedStreamExists?: Empty | undefined;
  expectedNoStream?: Empty | undefined;
}

export interface BatchAppendReq {
  correlationId: UUID | undefined;
  options: BatchAppendReq_Options | undefined;
  proposedMessages: BatchAppendReq_ProposedMessage[];
  isFinal: boolean;
}

export interface BatchAppendReq_Options {
  streamIdentifier: StreamIdentifier | undefined;
  streamPosition?: number | undefined;
  noStream?: Empty1 | undefined;
  any?: Empty1 | undefined;
  streamExists?: Empty1 | undefined;
  deadline21100?: Timestamp | undefined;
  deadline?: Duration | undefined;
}

export interface BatchAppendReq_ProposedMessage {
  id: UUID | undefined;
  metadata: { [key: string]: string };
  customMetadata: Uint8Array;
  data: Uint8Array;
}

export interface BatchAppendReq_ProposedMessage_MetadataEntry {
  key: string;
  value: string;
}

export interface BatchAppendResp {
  correlationId: UUID | undefined;
  error?: Status | undefined;
  success?: BatchAppendResp_Success | undefined;
  streamIdentifier: StreamIdentifier | undefined;
  streamPosition?: number | undefined;
  noStream?: Empty1 | undefined;
  any?: Empty1 | undefined;
  streamExists?: Empty1 | undefined;
}

export interface BatchAppendResp_Success {
  currentRevision?: number | undefined;
  noStream?: Empty1 | undefined;
  position?: AllStreamPosition | undefined;
  noPosition?: Empty1 | undefined;
}

export interface DeleteReq {
  options: DeleteReq_Options | undefined;
}

export interface DeleteReq_Options {
  streamIdentifier: StreamIdentifier | undefined;
  revision?: number | undefined;
  noStream?: Empty | undefined;
  any?: Empty | undefined;
  streamExists?: Empty | undefined;
}

export interface DeleteResp {
  position?: DeleteResp_Position | undefined;
  noPosition?: Empty | undefined;
}

export interface DeleteResp_Position {
  commitPosition: number;
  preparePosition: number;
}

export interface TombstoneReq {
  options: TombstoneReq_Options | undefined;
}

export interface TombstoneReq_Options {
  streamIdentifier: StreamIdentifier | undefined;
  revision?: number | undefined;
  noStream?: Empty | undefined;
  any?: Empty | undefined;
  streamExists?: Empty | undefined;
}

export interface TombstoneResp {
  position?: TombstoneResp_Position | undefined;
  noPosition?: Empty | undefined;
}

export interface TombstoneResp_Position {
  commitPosition: number;
  preparePosition: number;
}